pyproject.toml
[project]
name = "todo-fastapi-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt==4.0.1",
    "fastapi[all]>=0.135.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
